Add files via upload
The Dataset used was a headshot dataset of all celebrities. The query image will compare the face features to those of the celebrities and return the image of the celebrity your face is most similar to.
Histogram of oriented gradients (HOG) is a feature descriptor used to detect objects in computer vision and image processing. The HOG descriptor technique counts occurrences of gradient orientation in localized portions of an image - detection window, or region of interest (ROI).
Matlab Mapreduce programming is done to execute this project.
As Hadoop takes values in key value format, we provide the Map Reduce system with the image as the value and the its path as key. To establish this it is required that the image is stored for in a sequence file format and hence identity Mapper and Reducer have been used for this purpose. 
Computation of HOG Feature Vector 
    This steps involves computing the HOG feature vector of each image and storing it in another imagedatastore in Matlab.Further the Euclidian distance is found between the feature vector generated for the query image and every image in the dataset using Map Reduce. They are arranged in ascending order and the top 5 similar images are projected.For this two new sets of mappers and reducers are created.